Issue #1: Cannot execute WoK with gingers390x plug-in
the issue is because when wokd was run in development
environment, it failed to failed to initialize objectstore
since /var/lib/gingers390x/ folder was not created.
this patch includes changes in objectstore naming from
objectstore_gingers390x to objectstore since it is placed in
gingers390x folder.
Changes in Makefile.am, not to remove /var/lib/gingers390x/
folder on make uninstall.

Signed-off-by: Suresh Babu Angadi <sureshab@in.ibm.com>