Update Ginger s390x configuration according to recent changes on Wok
The application name and sub_nodes are restricted for the application
configuration and should never be edited. In that case, it is good to
remove this kind of information from a configuration file.

According to the recent changes on Wok, the application name will be the plugin
name with the first letter in uppercase. And the sub_nodes automatically
imported from control module.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>